+++ lib/intl_phone_picker/railtie.rb
@@ -1,7 +1,11 @@
 require 'intl_phone_picker/view_helpers'
+require 'intl_phone_picker/activerecord_helpers'
 
 module IntlPhonePicker
   class Engine < ::Rails::Engine
+    initializer "intl_phone_picker.activerecord_helpers" do
+      ActionView::Base.default_form_builder = IntlPhonePickerBuilder
+    end
   end
 
   class Railtie < Rails::Railtie