+++ lib/intl_phone_picker/view_helpers.rb
@@ -1,16 +1,16 @@
 module IntlPhonePicker
   module ViewHelpers
 
-    def intl_phone_fr_tag(name, value = nil, options = {})
-      options[:class] ||= String.new
-      options[:class] += ' intl_phone_input_fr'
-      telephone_field_tag(name = name, value = value, options = options)
+    def intl_phone_tag_fr(name, value = nil, options = {})
+      # options[:class] ||= String.new
+      # options[:class] += ' intl_phone_input_fr'
+      telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_fr'))
     end
 
-    def intl_phone_us_tag(name, value = nil, options = {})
-      options[:class] ||= String.new
-      options[:class] += ' intl_phone_input_us'
-      telephone_field_tag(name = name, value = value, options = options)
+    def intl_phone_tag_us(name, value = nil, options = {})
+      # options[:class] ||= String.new
+      # options[:class] += ' intl_phone_input_us'
+      telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_fr'))
     end
 
   end